+++ src/main/java/ua/knd11/controller/StudentController.java
@@ -2,8 +2,6 @@ package ua.knd11.controller;
 
 import ua.knd11.model.Student;
 import ua.knd11.model.User;
-import ua.knd11.security.AuthService;
-import ua.knd11.security.impl.AuthServiceImpl;
 import ua.knd11.service.StudentService;
 import ua.knd11.service.impl.StudentServiceImpl;
 
@@ -11,8 +9,6 @@ import java.util.List;
 
 public class StudentController {
     private final StudentService service = new StudentServiceImpl();
-    private final AuthService authService = new AuthServiceImpl();
-
     public void assignHeadStudent(int id) {
         service.assignHeadStudent(id);
     }
@@ -35,7 +31,6 @@ public class StudentController {
         try {
             Student student = new Student(surname, name, lastname, group, email, password);
             if (service.add(student)) {
-                authService.registration(student);
                 System.out.println("Student added successfully!");
             } else {
                 System.out.println("Student wasn't added.");
