+++ src/main/java/ua/knd11/model/Student.java
@@ -4,7 +4,7 @@ import ua.knd11.model.enums.StudentRole;
 
 import java.util.Locale;
 
-import static ua.knd11.service.impl.UserServiceImpl.normalizer;
+import static ua.knd11.util.FieldValidators.normalizer;
 
 public class Student extends User {
     private String group;
